colors script update, readme update

+++ scripts/discordColors.sh
@@ -1,27 +1,47 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 # RUN THIS SCRIPT RIGHT AFTER PYWAL TO PROPERLY APPLY COLORS!
 
 confPath="/home/$USER/.config"
 
-clientType="" # Vencord, Vesktop
-VENCORD="Vencord"
-VESKTOP="Vesktop"
+clientType=""
+supportedClients=("Vencord" "vesktop")
 
-if [[ "$confPath/$VENCORD" == *"cannot"* ]]; then 
-    if [[ "$confPath/$VESKTOP" == *"cannot"* ]]; then
-        echo "No compatible clients detected!"
-        echo "In order to use this script you have to use compatible discord clients! Check out github page."
-        echo "Exitting..."
-        exit
+function print_client_err() {
+    echo "No compatible clients detected!"
+echo "In order to use this script you have to use compatible discord clients! Check out github page for more info https://github.com/danihek/Themecord"
+    echo "Exitting..."
+}
+
+function print_dc_err() {
+    echo "Cannot find discord-colors.css in wal cache folder."
+    echo "Exitting..."
+}
+
+for client in "${supportedClients[@]}"; do
+    dc_path=~/.config/$client
+    if test -d $(realpath "$dc_path"); then
+        clientType="$client"
+        echo "[SUCCESS] $client DETECTED (path: $dc_path)"
+        break
     else
-        clientType=$VESKTOP
+        echo "[FAILURE] $client not DETECTED (path: $dc_path)"
+        continue
     fi
-else
-    clientType=$VENCORD
+done
+
+if [ -z "$clientType" ]; then
+    print_client_err
+    exit
 fi
 
 walColorsPath="/home/$USER/.cache/wal/colors-discord.css"
+
+if [ ! -f "$walColorsPath" ]; then
+    print_dc_err
+    exit
+fi
+
 themecordPath="/home/$USER/.config/$clientType/themes/Themecord.css"
 
 echo :root { > $themecordPath
